Fun with biblatex, an abstract, and download options
Tried to get biblatex working to fix issues with references, but I was
not smart enough. It's now working with natbib again. I added a draft
abstract (i.e., from the ReadMe file), and some wording about being able
to download the script and an example spawn database from GitHub to make
the script open and transparent.